requirements.txt
asgiref==3.11.1
boto3==1.42.88
botocore==1.42.88
crispy-bootstrap5==2026.3
Django==5.2.13
django-crispy-forms==2.6
django-extensions==4.1
django-mathfilters==1.0.0
django-storages==1.14.6
django-utils-six==2.0
jmespath==1.1.0
pillow==12.2.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
